♻️ tidy models
- move Roles & Permissions methods to own model file
- move tokens methods to own model file
- update imports
