+++ api/src/api/v1/prerequisites/__tests__/is_child_organisation.test.integration.ts
@@ -1,10 +1,10 @@
 import * as Hapi from 'hapi';
 import * as Knex from 'knex';
 import { init } from '../../../../../tests/utils/server';
-import { RoleEnum } from '../../../../auth/types';
 import { getConfig } from '../../../../../config';
 import { Users, CommunityBusinesses } from '../../../../models';
 import pre from '../is_child_organisation';
+import { RoleEnum } from '../../../../models/types';
 
 
 describe('Pre-requisite :: is_child_organisation', () => {
