+++ api/src/api/v1/prerequisites/__tests__/require_sibling_user.test.integration.ts
@@ -1,10 +1,10 @@
 import * as Hapi from 'hapi';
 import * as Knex from 'knex';
 import { init } from '../../../../../tests/utils/server';
-import { RoleEnum } from '../../../../auth/types';
 import { getConfig } from '../../../../../config';
 import { Users, Organisation, User, CommunityBusinesses, Volunteers } from '../../../../models';
 import pre from '../require_sibling_user';
+import { RoleEnum } from '../../../../models/types';
 
 
 describe('Pre-requisite :: is_sibling_user', () => {
