+++ api/src/api/v1/users/login.ts
@@ -6,8 +6,8 @@ import { Users, Organisations } from '../../../models';
 import { email, DEPRECATED_password, response } from './schema';
 import { Session, Token } from '../../../auth/strategies/standard';
 import { LoginRequest } from '../types';
-import { RoleEnum } from '../../../auth/types';
-import Roles from '../../../auth/roles';
+import Roles from '../../../models/role';
+import { RoleEnum } from '../../../models/types';
 
 
 const route: Hapi.ServerRoute[] = [
