+++ api/src/auth/index.ts
@@ -1,17 +1,10 @@
-import Permissions from './permissions';
-import Roles from './roles';
 import * as Standard from './strategies/standard';
 import * as External from './strategies/external';
-import { RoleEnum, PermissionLevelEnum } from './types';
-
+import { PermissionLevelEnum } from './types';
 
 const strategies = { Standard, External };
 
-
 export {
   PermissionLevelEnum,
-  RoleEnum,
-  Permissions,
-  Roles,
   strategies,
 };
